API: Add DynamicASN field for BGPPeers
Add the field to support FRR's automatic ASN detection
(https://docs.frrouting.org/en/latest/bgp.html#clicmd-neighbor-PEER-remote-as-internal).
Dynamic ASN is not an official name of this feature, but should be understandable.
Modifying the original ASN field to be IntOrString was considered but left out
since IntOrString is actually Int32OrString, and we explicitly need 32 bits (uint32) for the ASN.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>