E2E: increase debugging for the external peers
Currently we are not printing all debug messages for zebra and bfd
subsystem on the peers side in the e2etests. Therefore tests like
graceful restart test and future tests with BFD are harder to debug in
case of failure. This commit increases the debug in the frr config to be
the same as the frr config on the node, which includes zebra,bfd.

Signed-off-by: karampok <karampok@gmail.com>