Add redirect for the metallb.io/metallb vanity url
As part of the process of moving from metallb.universe.tf
to metallb.io, we'd have to rename the package.
In order to do so, a vanity url page must be implemented to refer
metallb.io/metallb to the actual repo.

Note: this could made parametric, and it will probably make sense to do
so if we need vanity urls for frr-k8s and / or a separated test repo.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>