E2E: extend frr provider interface to allow update
When testing backend, and the test fails in the after each we dump all
logs/config of all containers involved for debugging. When e2etest
includes speaker restart, and when we have frr-k8s backend this dump
fails because there is a static map between speaker pod and frrk8s
daemons. In this commit we extend the interface so test can call an
update function to update the map if they do a speaker restart.

Signed-off-by: karampok <karampok@gmail.com>