Support a configmap with node pattern when to stop advertising
Have a configMap which is mounted optionally in the speaker pod.
The content of the config map provides patterns like annotations, labels
which if matched on the node, advertisements stop instantly.
There is a loose match if any label or annotation from the running node
exists in the exclude pattern, then advertisments stop.

Signed-off-by: karampok <karampok@gmail.com>