IP Assignement: match clusterip family in case of single stack clusterip
There is no point in adding a dual stack loadbalancer ip if the cluster
ip is single family.

Additionally, when they do not match we feel free to clean the service
and reassign:
https://github.com/metallb/metallb/blob/v0.14.9/controller/service.go#L109

which might cause an active service to change its ip.

Also, refusing upfront a service requiring dual stack but coming with single (or
zero) clusterips.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>