FRR: deprecate disableMP
As described in metallb/metallb#2704, the genesis of the disableMP
flag is a bit unfortunate.

"When moving to FRR, we inadvertently enabled the possibility to
advertise the v6 addresses over a v4 session (and viceversa).
This was totally unintentional, we tried to amend and then we
found out that users (at least one!) were already relying on it.

This is causing confusion and works in a limited set of scenarios,
because FRR doesn't have a guarantee of what the v6 nexthop should
be for a session established over v6."

Because of this, we rollback the behavior, enabling an ipfamily to
a neighbor connected with the corresponding ip family.

We also add a flag "dualstackfamily" to enable the behavior for
those scenarios where a user might want to enable a dual stack
family for a neighbor.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>