FRR: refactor the prefix lists rendering logic
Instead of having a lot of logic in the template, we move it on the Go
code side. This has the advantage of having it more readeable and
maintainable.

The go code prepares the list of prefix lists related to communities,
local prefixes and prefixes to advertise.

Also, we select what to advertise depending on the ip family of the
neighbor.

Finally, we use two different prefix list names, one for ipv4 and one for
ivp6.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>