Add network policy support to Helm charts and manifests
- Network policies are disabled by default. User can enable them in Helm
  by setting the value `networkpolicies.enabled` true. When deploying
  using manifests the user should modify the config/native/kustomization.yaml
  to include the network policies.

- User can install a default deny network policy rule if needed. In Helm
  by setting the chart value `networkpolicies.defaultDeny` to true. In
  manifest by modifying accordingly the config/networkpolicies/kustomization.yaml
  to include or exclude the deny_all policy.

E2E tests are adding the network policies and the default deny rule
deny_all as part of the setup in the tasks.py file.

The Egress network policy port for the k8s API is 6443 because
the connection takes place through the k8s service, (as shown below).
This should be safe but in Helm charts there is a value to override it
if needed.

```
kubectl get svc kubernetes -o yaml
apiVersion: v1
kind: Service
metadata:
  name: kubernetes
  namespace: default
spec:
...
  ports:
  - name: https
    port: 443
    protocol: TCP
    targetPort: 6443
...
```

Signed-off-by: Konstantinos Karampogias <karampok@gmail.com>