E2E: refactor debug container executor and add go client implementation
I observed that the existing implementation works correctly for commands
that return instantly like whoami, ifconfig but no for commands that
take time to return and that they fail

e.g.
```
kubectl debug -it -n metallb-system -c controller-debugger-xxx
--target=controller --image=registry.k8s.io/e2e-test-images/agnhost:2.47
controller-765bbcb6fc-dzcv4 -- ./agnhost connect 10.244.1.7:8080 --timeout 5s

Targeting container "controller". If you don't see processes from this
container it may be because the container runtime doesn't support this
feature.
If you don't see a command prompt, try pressing enter.

 ---  hanging for a bit ---

TIMEOUT
Session ended, the ephemeral container will not be restarted but may be
reattached using 'kubectl attach controller-765bbcb6fc-dzcv4 -c
controller-debugger-xxx -i -t' if it is still running
```

For the commands that take time to return, current implementation is not
waiting and returns before the actual command finishes (probably due to
how TTY works), and therefore it returns neither the correct output (=
only the output from the command we want to run), neither the error (if
the command fails the kubectl debug actually returns ok).

So this commit refactors the code and follows the logic where the debug
container is a container attached to the pod, and we can do from there
normal `kubectl exec` commands without the `-it` argument and that
enables the e2etest for networkpolicies.

https://kubernetes.io/docs/tasks/debug/debug-application/debug-running-pod/#ephemeral-container

Signed-off-by: Konstantinos Karampogias <karampok@gmail.com>