E2E: make netpol tests infer controller ports dynamically
The existing implementation assumes that webhook is served by the
controller, but this is not the case when we deploy using the operator.
Then there is a dedicated webhook-only instance.

This change makes the test work in all deployment scenarios by
dynamically fetching the ports from the pod specification instead
of hardcoding them.

Signed-off-by: Konstantinos Karampogias <karampok@gmail.com>