config: validate duplicate BGP peers with node selector compatibility
Duplicate peers are permitted when they have node selectors that
either clearly don't overlap or when they have compatible
configurations that would result in identical BGP sessions.
Two peers are considered compatible when all their session
parameters match except for the node selectors themselves.

If it cannot definitively prove that node selectors are disjoint,
it requires the peers to be compatible.

Signed-off-by: Seena Fallah <seenafallah@gmail.com>