allocator: return error when pool has no available IPs
We explicitly return an error when there is no IPs in the pool
instead of relying on the caller to handle nil ips correctly.
As a side-effect, the AllocationFailed event changes from:
`Failed to allocate IP for "service-a": ["<nil>"] is not allowed in config for service`
to:
`Failed to allocate IP for "service-a": no available IPs in pool x`

When service-a requests an IP from a pool that doesn't have any left.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>